% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/france_7day_RTO.R
\name{fr_7forecast_dw}
\alias{fr_7forecast_dw}
\title{7day RTE-France forecast download}
\usage{
fr_7forecast_dw(ddate, hour.freq = T)
}
\arguments{
\item{ddate}{a date object with ymd format}

\item{hour.freq}{a logical object for getting data in hourly}
}
\value{
a dataframe with the scraped forecast data
}
\description{
7day RTE-France forecast download
}
\examples{
fr_7forecast_dw(Sys.Date())
fr_7forecast_dw("2017-09-25", hour.freq=F )
}
