% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fr.R
\name{fr_load_dw}
\alias{fr_load_dw}
\title{fr_load_dw}
\usage{
fr_load_dw(dag, hour.freq = T)
}
\arguments{
\item{dag}{a date object. this is the date for which the load data is
downloaded}

\item{hour.freq}{a logical object. this is to get hourly frequency if true}
}
\value{
a data.frame that has date and Load (numeric) data
}
\description{
This function returns the load (Electricity Demand) at a particular day in
France. Data Source is RTE
}
\examples{
fr_load_dw (Sys.Date(), hour.freq=T)
fr_load_dw ("2017-01-01") #date format is ymd

}
