% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjm_7day.R
\name{pjm_7dag_forecast}
\alias{pjm_7dag_forecast}
\title{PJM seven day load forecast. Scraped from PJM website}
\usage{
pjm_7dag_forecast()
}
\value{
a data frame which has the current pjm 7 day load foreacst
}
\description{
PJM seven day load forecast. Scraped from PJM website
}
\examples{
pjm_7dag_forecast() # no argumets required.. this data is available only for today
}
